Module Client.Stream

Control for a polling loop

type 'a t

val cancel : _ t -> unit Lwt.t

cancel t notify the server that we are stopping our polling loop. It is an error to call next after cancel

val next : 'a t -> 'a option Lwt.t

next t poll for the next value. It is an error to call next again until the previous next terminated. If next returns None, subsequent calls to next is forbidden.